[LibOS,PAL] Remove `struct pal_iovec`
This struct was essentially `struct iovec` from Linux and it's usage
required some otherwise unneeded `malloc()` + `free()`, which became
a performance bottleneck in some cases.
