+++ common/include/iovec.h
@@ -1,0 +1,11 @@
+/* SPDX-License-Identifier: LGPL-3.0-or-later */
+/* Copyright (C) 2022 Intel Corporation
+ *                    Borys Popławski <borysp@invisiblethingslab.com>
+ */
+
+#pragma once
+
+struct iovec {
+    void* iov_base;
+    size_t iov_len;
+};
