+++ pal/src/host/linux-sgx/pal_linux_types.h
@@ -40,23 +40,3 @@ struct sockaddr {
     unsigned short sa_family;
     char sa_data[128 - sizeof(unsigned short)];
 };
-
-struct cmsghdr {
-    size_t cmsg_len;
-    int cmsg_level;
-    int cmsg_type;
-};
-
-#ifndef SCM_RIGHTS
-#define SCM_RIGHTS 1
-#endif
-
-#define CMSG_DATA(cmsg)         ((unsigned char*)((struct cmsghdr*)(cmsg) + 1))
-#define CMSG_NXTHDR(mhdr, cmsg) __cmsg_nxthdr(mhdr, cmsg)
-#define CMSG_FIRSTHDR(mhdr)                                   \
-    ((size_t)(mhdr)->msg_controllen >= sizeof(struct cmsghdr) \
-         ? (struct cmsghdr*)(mhdr)->msg_control               \
-         : (struct cmsghdr*)0)
-#define CMSG_ALIGN(len) ALIGN_UP(len, sizeof(size_t))
-#define CMSG_SPACE(len) (CMSG_ALIGN(len) + CMSG_ALIGN(sizeof(struct cmsghdr)))
-#define CMSG_LEN(len)   (CMSG_ALIGN(sizeof(struct cmsghdr)) + (len))
