+++ pal/src/host/linux-sgx/pal_sockets.c
@@ -454,7 +454,7 @@ static int attrsetbyhdl_udp(PAL_HANDLE handle, PAL_STREAM_ATTR* attr) {
     return attrsetbyhdl_common(handle, attr);
 }
 
-static int send(PAL_HANDLE handle, struct pal_iovec* pal_iov, size_t iov_len, size_t* out_size,
+static int send(PAL_HANDLE handle, struct iovec* iov, size_t iov_len, size_t* out_size,
                 struct pal_socket_addr* addr, bool force_nonblocking) {
     assert(handle->hdr.type == PAL_TYPE_SOCKET);
 
@@ -468,19 +468,9 @@ static int send(PAL_HANDLE handle, struct pal_iovec* pal_iov, size_t iov_len, si
         assert(linux_addrlen <= INT_MAX);
     }
 
-    struct iovec* iov = malloc(iov_len * sizeof(*iov));
-    if (!iov) {
-        return -PAL_ERROR_NOMEM;
-    }
-    for (size_t i = 0; i < iov_len; i++) {
-        iov[i].iov_base = pal_iov[i].iov_base;
-        iov[i].iov_len = pal_iov[i].iov_len;
-    }
-
     unsigned int flags = force_nonblocking ? MSG_DONTWAIT : 0;
     ssize_t ret = ocall_send(handle->sock.fd, iov, iov_len, addr ? &sa_storage : NULL,
                              linux_addrlen, /*control=*/NULL, /*controllen=*/0, flags);
-    free(iov);
     if (ret < 0) {
         return unix_to_pal_error(ret);
     }
@@ -488,21 +478,12 @@ static int send(PAL_HANDLE handle, struct pal_iovec* pal_iov, size_t iov_len, si
     return 0;
 }
 
-static int recv(PAL_HANDLE handle, struct pal_iovec* pal_iov, size_t iov_len,
-                size_t* out_total_size, struct pal_socket_addr* addr, bool force_nonblocking) {
+static int recv(PAL_HANDLE handle, struct iovec* iov, size_t iov_len, size_t* out_total_size,
+                struct pal_socket_addr* addr, bool force_nonblocking) {
     assert(handle->hdr.type == PAL_TYPE_SOCKET);
 
     struct sockaddr_storage sa_storage;
     size_t linux_addrlen = addr ? sizeof(sa_storage) : 0;
-
-    struct iovec* iov = malloc(iov_len * sizeof(*iov));
-    if (!iov) {
-        return -PAL_ERROR_NOMEM;
-    }
-    for (size_t i = 0; i < iov_len; i++) {
-        iov[i].iov_base = pal_iov[i].iov_base;
-        iov[i].iov_len = pal_iov[i].iov_len;
-    }
 
     unsigned int flags = force_nonblocking ? MSG_DONTWAIT : 0;
     if (handle->sock.type == PAL_SOCKET_UDP) {
@@ -512,7 +493,6 @@ static int recv(PAL_HANDLE handle, struct pal_iovec* pal_iov, size_t iov_len,
     }
     ssize_t ret = ocall_recv(handle->sock.fd, iov, iov_len, addr ? &sa_storage : NULL,
                              &linux_addrlen, /*control=*/NULL, /*controllenptr=*/NULL, flags);
-    free(iov);
     if (ret < 0) {
         return unix_to_pal_error(ret);
     }
@@ -633,7 +613,7 @@ int _PalSocketConnect(PAL_HANDLE handle, struct pal_socket_addr* addr,
     return handle->sock.ops->connect(handle, addr, out_local_addr);
 }
 
-int _PalSocketSend(PAL_HANDLE handle, struct pal_iovec* iov, size_t iov_len, size_t* out_size,
+int _PalSocketSend(PAL_HANDLE handle, struct iovec* iov, size_t iov_len, size_t* out_size,
                    struct pal_socket_addr* addr, bool force_nonblocking) {
     if (!handle->sock.ops->send) {
         return -PAL_ERROR_NOTSUPPORT;
@@ -641,7 +621,7 @@ int _PalSocketSend(PAL_HANDLE handle, struct pal_iovec* iov, size_t iov_len, siz
     return handle->sock.ops->send(handle, iov, iov_len, out_size, addr, force_nonblocking);
 }
 
-int _PalSocketRecv(PAL_HANDLE handle, struct pal_iovec* iov, size_t iov_len, size_t* out_total_size,
+int _PalSocketRecv(PAL_HANDLE handle, struct iovec* iov, size_t iov_len, size_t* out_total_size,
                    struct pal_socket_addr* addr, bool force_nonblocking) {
     if (!handle->sock.ops->recv) {
         return -PAL_ERROR_NOTSUPPORT;
