+++ pal/src/host/linux/pal_streams.c
@@ -6,11 +6,11 @@
  */
 
 #include <asm/fcntl.h>
-#include <netinet/in.h>
 #include <stdalign.h>
 #include <stdbool.h>
 
 #include "api.h"
+#include "linux_socket.h"
 #include "linux_utils.h"
 #include "pal.h"
 #include "pal_error.h"
